ParfBar/requirements.txt
aiogram>=3.22.0
SQLAlchemy>=2.0.44
alembic>=1.17.2
aiofiles==24.1.0
python-dotenv>=1.2.1
